examples/django-drf/pyproject.toml
[project]
name = "paygate-example-django"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "django>=5.0",
  "djangorestframework>=3.15",
  "paygate[django]",
  "gunicorn>=22",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
